split task/service filters
