+++ lib/process-services-cloud/src/lib/task/task-filters/components/edit-task-filters/edit-service-task-filter/edit-service-task-filter-cloud.component.spec.ts
@@ -19,21 +19,21 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { SimpleChange } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { of, Subject } from 'rxjs';
-import { TASK_FILTERS_SERVICE_TOKEN } from '../../../../services/cloud-token.service';
-import { LocalPreferenceCloudService } from '../../../../services/local-preference-cloud.service';
-import { ProcessServiceCloudTestingModule } from '../../../../testing/process-service-cloud.testing.module';
-import { AppsProcessCloudService } from '../../../../app/services/apps-process-cloud.service';
-import { fakeApplicationInstance, fakeApplicationInstanceWithEnvironment } from '../../../../app/mock/app-model.mock';
-import { TaskFiltersCloudModule } from '../../task-filters-cloud.module';
-import { ServiceTaskFilterCloudService } from '../../services/service-task-filter-cloud.service';
-import { TaskCloudService } from '../../../services/task-cloud.service';
-import { fakeServiceFilter } from '../../mock/task-filters-cloud.mock';
+import { TASK_FILTERS_SERVICE_TOKEN } from '../../../../../services/cloud-token.service';
+import { LocalPreferenceCloudService } from '../../../../../services/local-preference-cloud.service';
+import { ProcessServiceCloudTestingModule } from '../../../../../testing/process-service-cloud.testing.module';
+import { AppsProcessCloudService } from '../../../../../app/services/apps-process-cloud.service';
+import { fakeApplicationInstance, fakeApplicationInstanceWithEnvironment } from '../../../../../app/mock/app-model.mock';
+import { TaskFiltersCloudModule } from '../../../task-filters-cloud.module';
+import { ServiceTaskFilterCloudService } from '../../../services/service-task-filter-cloud.service';
+import { TaskCloudService } from '../../../../services/task-cloud.service';
+import { fakeServiceFilter } from '../../../mock/task-filters-cloud.mock';
 import { EditServiceTaskFilterCloudComponent } from './edit-service-task-filter-cloud.component';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
-import { ProcessDefinitionCloud } from '../../../../models/process-definition-cloud.model';
-import { TaskFilterDialogCloudComponent } from '../task-filter-dialog/task-filter-dialog-cloud.component';
-import { fakeEnvironmentList } from '../../../../common/mock/environment.mock';
-import { mockApplicationTaskFilterProperties } from '../../mock/edit-task-filter-cloud.mock';
+import { ProcessDefinitionCloud } from '../../../../../models/process-definition-cloud.model';
+import { TaskFilterDialogCloudComponent } from '../../task-filter-dialog/task-filter-dialog-cloud.component';
+import { fakeEnvironmentList } from '../../../../../common/mock/environment.mock';
+import { mockApplicationTaskFilterProperties } from '../../../mock/edit-task-filter-cloud.mock';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { MatExpansionPanelHarness } from '@angular/material/expansion/testing';
