+++ lib/process-services-cloud/src/lib/task/task-filters/components/service-task-filters/service-task-filters-cloud.component.spec.ts
@@ -18,13 +18,13 @@
 import { SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { of, throwError } from 'rxjs';
-import { TASK_FILTERS_SERVICE_TOKEN } from '../../../services/cloud-token.service';
-import { LocalPreferenceCloudService } from '../../../services/local-preference-cloud.service';
+import { TASK_FILTERS_SERVICE_TOKEN } from '../../../../services/cloud-token.service';
+import { LocalPreferenceCloudService } from '../../../../services/local-preference-cloud.service';
 import { By } from '@angular/platform-browser';
-import { ProcessServiceCloudTestingModule } from '../../../testing/process-service-cloud.testing.module';
-import { TaskFiltersCloudModule } from '../task-filters-cloud.module';
-import { fakeGlobalServiceFilters } from '../mock/task-filters-cloud.mock';
-import { ServiceTaskFilterCloudService } from '../services/service-task-filter-cloud.service';
+import { ProcessServiceCloudTestingModule } from '../../../../testing/process-service-cloud.testing.module';
+import { TaskFiltersCloudModule } from '../../task-filters-cloud.module';
+import { fakeGlobalServiceFilters } from '../../mock/task-filters-cloud.mock';
+import { ServiceTaskFilterCloudService } from '../../services/service-task-filter-cloud.service';
 import { ServiceTaskFiltersCloudComponent } from './service-task-filters-cloud.component';
 
 describe('ServiceTaskFiltersCloudComponent', () => {
