+++ lib/process-services-cloud/src/lib/task/task-filters/components/task-filters/task-filters-cloud.component.scss
@@ -1,0 +1,39 @@
+.adf-task-filters {
+    margin-right: calc(-1 * var(--adf-theme-spacing));
+
+    &__entry {
+        font-size: var(--theme-body-1-font-size);
+        color: var(--adf-theme-foreground-text-color-054);
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        flex: 1;
+        height: 100%;
+
+        &:hover {
+            color: var(--theme-primary-color);
+        }
+    }
+
+    &__entry-label {
+        display: flex;
+        flex: 1;
+        align-items: center;
+        gap: var(--adf-theme-spacing);
+    }
+
+    &__entry-counter {
+        padding: 0 5px;
+        border-radius: 15px;
+
+        &.adf-active {
+            background-color: var(--theme-accent-color);
+            color: var(--theme-accent-color-default-contrast);
+            font-size: smaller;
+        }
+    }
+
+    .adf-active .adf-task-filters__entry-label {
+        color: var(--theme-primary-color);
+    }
+}
