+++ lib/process-services-cloud/src/lib/task/task-filters/components/task-filters/task-filters-cloud.component.ts
@@ -17,21 +17,21 @@
 
 import { Component, EventEmitter, inject, Input, OnChanges, OnInit, Output, SimpleChanges, ViewEncapsulation } from '@angular/core';
 import { Observable } from 'rxjs';
-import { TaskFilterCloudService } from '../services/task-filter-cloud.service';
-import { FilterParamsModel, TaskFilterCloudModel } from '../models/filter-cloud.model';
+import { TaskFilterCloudService } from '../../services/task-filter-cloud.service';
+import { FilterParamsModel, TaskFilterCloudModel } from '../../models/filter-cloud.model';
 import { AppConfigService, TranslationService } from '@alfresco/adf-core';
 import { debounceTime, tap } from 'rxjs/operators';
-import { BaseTaskFiltersCloudComponent } from './base-task-filters-cloud.component';
-import { TaskDetailsCloudModel } from '../../models/task-details-cloud.model';
-import { TaskCloudEngineEvent } from '../../../models/engine-event-cloud.model';
-import { TaskListCloudService } from '../../task-list/services/task-list-cloud.service';
-import { TaskFilterCloudAdapter } from '../../../models/filter-cloud-model';
+import { BaseTaskFiltersCloudComponent } from '../base-task-filters-cloud.component';
+import { TaskDetailsCloudModel } from '../../../models/task-details-cloud.model';
+import { TaskCloudEngineEvent } from '../../../../models/engine-event-cloud.model';
+import { TaskListCloudService } from '../../../task-list/services/task-list-cloud.service';
+import { TaskFilterCloudAdapter } from '../../../../models/filter-cloud-model';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Component({
     selector: 'adf-cloud-task-filters',
-    templateUrl: './base-task-filters-cloud.component.html',
-    styleUrls: ['./base-task-filters-cloud.component.scss'],
+    templateUrl: './task-filters-cloud.component.html',
+    styleUrls: ['./task-filters-cloud.component.scss'],
     encapsulation: ViewEncapsulation.None
 })
 export class TaskFiltersCloudComponent extends BaseTaskFiltersCloudComponent implements OnInit, OnChanges {
